ETL/senegal_export/requirements.txt
dbt-postgres
pandas
sqlalchemy
psycopg2-binary
subprocess
time

